conf/requirements.txt
allianceauth-mumbletemps
django-eveuniverse
aa-structures
superlance